requirements.txt
click==8.1.7
rich==13.7.0
twilio==8.11.0
requests==2.31.0
python-dotenv==1.0.0
pyyaml==6.0.1
sqlalchemy==2.0.23
cryptography==41.0.7
phonenumbers==8.13.26
